Server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
aiosmtplib==2.0.2
    # via fastapi-mail
alt-profanity-check==1.5.2
    # via -r .\requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   asyncer
    #   httpx
    #   starlette
    #   watchfiles
apscheduler==3.10.4
    # via -r .\requirements.in
asyncer==0.0.8
    # via fastapi-cdn-host
bcrypt==4.2.0
    # via -r .\requirements.in
beautifulsoup4==4.12.3
    # via -r .\requirements.in
blinker==1.8.2
    # via fastapi-mail
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==43.0.1
    # via -r .\requirements.in
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   fastapi-mail
fastapi[standard]==0.115.0
    # via
    #   -r .\requirements.in
    #   fastapi-cache2
    #   fastapi-cdn-host
fastapi-cache2==0.2.2
    # via -r .\requirements.in
fastapi-cdn-host==0.8.0
    # via -r .\requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
fastapi-mail==1.4.1
    # via -r .\requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via
    #   fastapi
    #   fastapi-cdn-host
icalendar==6.0.0
    # via -r .\requirements.in
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   fastapi
    #   fastapi-mail
joblib==1.4.2
    # via
    #   alt-profanity-check
    #   scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.1.1
    # via
    #   scikit-learn
    #   scipy
pendulum==3.0.0
    # via fastapi-cache2
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   fastapi
    #   fastapi-mail
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.5.2
    # via fastapi-mail
pygments==2.18.0
    # via rich
pyjwt==2.9.0
    # via -r .\requirements.in
pymysql==1.1.1
    # via -r .\requirements.in
pyotp==2.9.0
    # via -r .\requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r .\requirements.in
    #   icalendar
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.12
    # via fastapi
pytz==2024.2
    # via
    #   -r .\requirements.in
    #   apscheduler
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via -r .\requirements.in
rich==13.8.1
    # via
    #   -r .\requirements.in
    #   typer
scikit-learn==1.5.2
    # via alt-profanity-check
scipy==1.14.1
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   apscheduler
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.35
    # via -r .\requirements.in
starlette==0.38.6
    # via
    #   fastapi
    #   fastapi-mail
threadpoolctl==3.5.0
    # via scikit-learn
time-machine==2.16.0
    # via pendulum
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   fastapi-cache2
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
tzdata==2024.2
    # via
    #   icalendar
    #   pendulum
    #   tzlocal
tzlocal==5.2
    # via apscheduler
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.31.0
    # via
    #   -r .\requirements.in
    #   fastapi
    #   fastapi-cache2
    #   fastapi-cli
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
